+++ src/app/explore/page.tsx
@@ -6,9 +6,8 @@ import { useUserContext } from '../userContext'
 import { create_user, get_user } from '@/lib/prismaUtils'
 import { generateFromEmail } from 'unique-username-generator'
 import { AvatarGenerator } from 'random-avatar-generator'
-import { AppName } from '@config/app'
-import particle from '@/lib/particle'
 import { GiDigitalTrace } from "react-icons/gi"
+import particle from '@config/authConfig'
 
 const Page = () => {
   const [view, setView] = useState<boolean>(false)
